requirements.txt
discord.py==2.0.0
luawl.py==1.2.0
